tests/dup3.c: print syscall() result using %ld format

+++ tests/dup3.c
@@ -13,10 +13,10 @@ main(void)
 {
 	const long int fd_old = (long int) 0xdeadbeefffffffff;
 	const long int fd_new = (long int) 0xdeadbeeffffffffe;
-	int rc = syscall(__NR_dup3, fd_old, fd_new, O_CLOEXEC);
-	printf("dup3(%d, %d, O_CLOEXEC) = %d %s (%m)\n",
-	       (int) fd_old, (int) fd_new, rc,
-	       errno2name());
+
+	long rc = syscall(__NR_dup3, fd_old, fd_new, O_CLOEXEC);
+	printf("dup3(%d, %d, O_CLOEXEC) = %ld %s (%m)\n",
+	       (int) fd_old, (int) fd_new, rc, errno2name());
 
 	puts("+++ exited with 0 +++");
 	return 0;
